Changed files so that TimingSequence is a field in RemoteControl
In order to make the interface easier to use for single-shot
measurements, I have added a field 'sq' to the RemoteControl
class.  I have also added a 'makerCallback' property and a
'make' function.  For a RemoteControl object r, a call of

r.make(varargin)

will call the 'makerCallback' function with the varargin{:}
arguments.  If no callback is supplied, it uses the makeSequence
function that should be found in the same folder.  This means that
if one wanted to run a sequence in a single line, one would use

r.make(varargin).upload.run;

I have additionally set up the makeSequence function to automatically
upload and run the sequence that is created by that function if no
output argument is requested.  Therefore, just running the function
using F5 or whatever will upload and run that sequence.  I have not
tested this, and I am unsure how garbage collection will work as the
RemoteControl variable is not explicitly released.