Changes to sequence, added image sequence function
Some changes to the sequence -- current values create a BEC but
apply a repump pulse 20 ms before imaging so that the cloud expands
more and thus the number of atoms can be determined with on-resonance
imaging.

Added a function to makeSequenceFull that creates a set of commands
for imaging.  It has multiple parameters for creating different
timings for imaging.

Changed the RemoteControl function so that the setFunc() function
is in init() and not run().  This allows me to set the
BytesAvailableFcn for the conn property to Abs_Analysis('last')
which means that when I run a single sequence from the command
line it automatically analyzes the absorption image.