Significant changes for CTD apparatus
Added a DDSChannel class to handle DDS channels that are closely
integrated into the control interface.  This is handled mostly by
changing the TimingControllerChannel class to handle value arrays
that are 2D, with rows being different time entries and columns
being different "sub-values".  Currently only the DDSChannel class
uses this feature with the value being [frequency,amplitude,phase]

DDSChannel has its own compile function, and compiling and uploading
data is handled separately in the RemoteControl class.

No testing has been done to see if it works besides basic tests
of the DDSChannel class and the new sub-valued value entries

initSequence() has names that correspond to CTD ports, but defaults
are not set yet and makeSequence is still the old gravimeter
sequence