Modified sequence creation so that DDS pulses are handled internally
Previously, DDS control was handled exclusively by a separate set
of classes.  It is better for all experimental control to be
handled by the same set of classes.  This update partially
implements DDS control exclusively through the gravimeter interface.

The two DDS channels on the MOGLabs box are implemented as channel
objects in the TimingSequence.  These channels have value arrays
that are Nx3 where N is the number of updates and the columns are
frequency, relative optical power, and phase.  This values can
be handled by the normal TimingControllerChannel methods.

Two additional stand-alone functions have been added that condense
creating a sequence of Bragg pulses or a Raman pulse into one
function.  These are called with the first argument being the array
of DDS channels, and the next arguments being name/value pairs
specifying the behaviour of the pulse/pulse sequence.

Uploading data to the MOGLabs box still uses the classes in
interfaces/moglabs-arf-xrf.  Specifically, the values compiled
and returned by the DDSChannel class are fed into instances of
the mogtable class, which has as a parent a mogdevice object.
Reduction of the number of points, as well as synchronization
between the two channels, is handled by the mogtable class.
Uploading is done asynchronously and is also handled by the
mogtable and mogdevice classes.