Changed how listeners are handled and improved plot efficiency
Had an issue with updating the plot when using the DDS channels
as they have lots of points.  Repeated calls to
TimingSequence.latest slowed things down enormously.  This was
easily solved by calling it once outside of the plotting loop
and then using that value in the loop.

Also created a delete method for the listener that listens for
the change in the sequence.