Minor changes to uploading procedure for DDS
Table synchronization between the DDS channels can now take place
with either channel 1 or channel 2 being the master.  Channel 1
is by default the master channel which provides synchronization
information to channel 2, but if there are no points in channel 1
then channel 2 becomes the master.  From the point of view of the
MOGLabs ARF box, channel 1 is always the master, it is just a
question of which set of commands is used to ensure that both
channels have updates at the same time.

Also changed the upload process so that if the time it takes to
upload instructions to the MOGLabs ARF box is longer than the time
to the DDS trigger, then the uploading process is paused for a
certain amount of time to make sure that the upload completes
before the box is triggered.