Added saveSequenceCopy() which creates a copy of makeSequence
The function creates a copy of the current makeSequence file that
is being used, and then adds additional lines that show what the
input arguments were when the makeSequence file was run.  This
should allow the function to be run without any knowledge of what
the input arguments were.

Additionally, the saved files are now organized into folders by
year, month, and day