Updated RemoteControl + DDSChannel for easy switching between dBm and hex
DDSChannel now has a property 'power_conversion_method' for controlling
how the normalized power set in the sequence is converted into units the
DDS understands.  This property is stored in the compiled data, and that
field in the data structure is used to determine how the DDS data is
uploaded to the MOGLabs ARF unit.  initSequence.m has been updated so
that the DDS data is uploaded via binary upload.

Some minor updates to makeBEC.m to reduce the number of samples.