Major changes to sequence options, updated make functions to match
Sequence options are now instances of the abstract SequenceOptionsAbstract
class, which has set() and replace() functions that the sub-classes
inherit.  This allows for better categorization of options.  Structure is

SequenceOptions
  - BraggSequenceOptions
  - RamanSequenceOptions
  - MicrowaveSequenceOptions

and each instance has its own properties.

Also changed the sequence to interface better with the LabVIEW camera
system.  Now there is a camDelay property that needs to be set and then
written to the LabVIEW controller.  This is used to trigger the
camera acquisition system.  Generally set it about 2 s before the
images are to be taken.

I have also created a class called RunConversions that has all of the
parameter conversions for different channels.  Use these to convert
between semantic values (G, G/cm, MHz, etc.) to actual Volts.