Added triggering ability for camera VI, sequence updates
Camera VI is now triggered from the LabVIEW control interface, and
this needs a delay between when the sequence starts and when the
camera should be triggered to avoid long timeouts.  The camDelay property
has been added to TimingSequence.m and an associated upload word
to RemoteControl.

Bug fix to SequenceOptionsAbstract.m

Added a 4-image imaging sequence function