Changed TCP/IP functions as tcpip() no longer exists in MATLAB 2022. In progress.
Previously TCP/IP client connections were handled with tcpip('networkrole','client')
but that function, as well as a lot of the instrument control toolbox,
has been removed.  Client connections are now handled using
tcpclient().  This has some differences in how data is read/written.

Reading lines of text is done through readline(), and writing through
writeline() rather than fgetl() and fwrite() or fprintf().

On the upside, it looks like instrument objects are deleted the moment
that the object itself is deleted, so there should be no more hidden
objects.